MarkAsDelivered method... needs improvements
MarkAsDelivered has 1. auto set receivedate, 2. it is supposed to get employeeId from auth but... now its getting it from url 3.updates/adds to customer balance if (price > 0) 4. shipment is delivered so delivery is supposed to be paid - set to true if false. ALL these changes made to shipment. PLUS it makes record in shipmentstatushistory to "DELIVERED"