Update JwtService.java to include user roles in the bearer token
I removed the getSignInKey method and created a key field that is initialized once. This way, you don’t need to decode and create the key every time you need to use it.

In the generateToken method that takes a UserDetails parameter, I cast UserDetails to User and get the user’s roles. I then put the roles in the claims map. This way, the user’s roles are included in the JWT.

I modified the generateToken method that takes a claims map and a subject string. Now, this method doesn’t need to know about UserDetails. It just needs the claims and the subject, which makes it more reusable.